Reflexion changes TSOM
