+++ agential/cog/agent/.plandex/project.json
@@ -1,0 +1,1 @@
+{"id":"d097a18f-46aa-46e5-914b-e4a7517a9df8"}
