--- agential/cog/modules/plan/__init__.py
@@ -1,1 +1,0 @@
-"""Planning modules."""
