--- agential/cog/modules/plan/base.py
@@ -1,1 +1,0 @@
-"""Base planning module."""
