--- agential/cog/modules/reflect/__init__.py
@@ -1,1 +1,0 @@
-"""Reflecting modules."""
