--- agential/cog/modules/score/__init__.py
@@ -1,1 +1,0 @@
-"""Scoring modules."""
