+++ agential/cog/reflexion/agent.py
@@ -13,11 +13,11 @@ from typing import Any, Dict, List, Optional, Tuple
 from langchain_core.language_models.chat_models import BaseChatModel
 
 from agential.base.agent import BaseAgent
-from agential.cog.modules.reflect.reflexion import (
+from agential.cog.reflexion.reflect import (
     ReflexionCoTReflector,
     ReflexionReActReflector,
 )
-from agential.cog.output.reflexion import (
+from agential.cog.reflexion.output import (
     ReflexionCoTOutput,
     ReflexionReActOutput,
     ReflexionReActStepOutput,
