+++ agential/cog/reflexion/strategies/base.py
@@ -5,7 +5,7 @@ from typing import Any, Dict, List, Optional, Tuple
 
 from langchain_core.language_models.chat_models import BaseChatModel
 
-from agential.cog.output.reflexion import ReflexionReActStepOutput
+from agential.cog.reflexion.output import ReflexionReActStepOutput
 from agential.base.strategies import BaseStrategy
