+++ agential/cog/reflexion/strategies/code.py
@@ -1,4 +1,4 @@
-"""Reflexion Agent strategies for QA."""
+"""Reflexion Agent strategies for Code."""
 
 import re
 
@@ -6,55 +6,87 @@ from typing import Any, Dict, List, Optional, Tuple
 
 import tiktoken
 
-from langchain_community.docstore.wikipedia import Wikipedia
 from langchain_core.language_models.chat_models import BaseChatModel
-from tiktoken import Encoding
+from tiktoken.core import Encoding
 
 from agential.cog.eval.em import EM
-from agential.cog.functional.reflexion import (
+from agential.cog.reflexion.functional import (
     _is_halted,
     _prompt_cot_agent,
     _prompt_react_agent,
     _truncate_scratchpad,
 )
-from agential.cog.modules.reflect.reflexion import (
+from agential.cog.reflexion.reflect import (
     ReflexionCoTReflector,
     ReflexionReActReflector,
 )
-from agential.cog.output.reflexion import ReflexionReActStepOutput
-from agential.cog.strategies.reflexion.base import (
+from agential.cog.reflexion.output import ReflexionReActStepOutput
+from agential.cog.reflexion.strategies.base import (
     ReflexionCoTBaseStrategy,
     ReflexionReActBaseStrategy,
 )
-from agential.utils.docstore import DocstoreExplorer
+from agential.utils.general import safe_execute
 from agential.utils.parse import remove_newline
 
 
-def parse_qa_action(string: str) -> Tuple[str, str]:
-    """Parses an action string into an action type and its argument.
+def parse_code_action_cot(action: str) -> Tuple[str, str]:
+    """Parses an action string to extract the action type and code content.
 
-    This method is used in ReAct and Reflexion.
+    Identifies action types (`Finish`) and extracts the
+    corresponding code content enclosed within Markdown-style code blocks.
+    The action type is case-insensitive and the code content is trimmed of
+    leading and trailing whitespace.
 
     Args:
-        string (str): The action string to be parsed.
+        action (str): The action string containing the action type and code content.
 
     Returns:
-        Tuple[str, str]: A tuple containing the action type and argument.
+        Tuple[str, str]: A tuple containing the extracted action type (capitalized)
+        and the extracted code content.
     """
-    pattern = r"^(\w+)\[(.+)\]$"
-    match = re.match(pattern, string)
+    action_split = action.split("```python", maxsplit=1)
+    match = re.search(r"\b(Finish)\b", action_split[0], re.IGNORECASE)
 
-    if match:
-        action_type = match.group(1)
-        argument = match.group(2)
-    else:
+    action_type = match.group(0).lower().capitalize() if match else ""
+    try:
+        query = action_split[1].split("```")[0].strip() if action_type else ""
+    except:
         action_type = ""
-        argument = ""
-    return action_type, argument
+        query = ""
+
+    return action_type, query
 
 
-class ReflexionCoTQAStrategy(ReflexionCoTBaseStrategy):
-    """A strategy class for QA benchmarks using the ReflexionCoT agent.
+def parse_code_action_react(action: str) -> Tuple[str, str]:
+    """Parses an action string to extract the action type and code content.
+
+    Identifies action types (`Finish`, `Implement`, `Test`) and extracts the
+    corresponding code content enclosed within Markdown-style code blocks.
+    The action type is case-insensitive and the code content is trimmed of
+    leading and trailing whitespace.
+
+    Args:
+        action (str): The action string containing the action type and code content.
+
+    Returns:
+        Tuple[str, str]: A tuple containing the extracted action type (capitalized)
+        and the extracted code content.
+    """
+    action_split = action.split("```python", maxsplit=1)
+    match = re.search(r"\b(Finish|Test|Implement)\b", action_split[0], re.IGNORECASE)
+
+    action_type = match.group(0).lower().capitalize() if match else ""
+    try:
+        query = action_split[1].split("```")[0].strip() if action_type else ""
+    except:
+        action_type = ""
+        query = ""
+
+    return action_type, query
+
+
+class ReflexionCoTCodeStrategy(ReflexionCoTBaseStrategy):
+    """A strategy class for Code benchmarks using the ReflexionCoT agent.
 
     Attributes:
         llm (BaseChatModel): The language model used for generating answers and critiques.
@@ -153,9 +185,10 @@ class ReflexionCoTQAStrategy(ReflexionCoTBaseStrategy):
             prompt=prompt,
             additional_keys=additional_keys,
         )
-        action = remove_newline(action).strip()
-        self._scratchpad += " " + action
-        action_type, query = parse_qa_action(action)
+        action = action.split("Observation")[0].strip()
+
+        action_type, query = parse_code_action_cot(action)
+        self._scratchpad += f" {action_type}[\n```python\n{query}\n```\n]"
 
         return action_type, query
 
@@ -172,19 +205,22 @@ class ReflexionCoTQAStrategy(ReflexionCoTBaseStrategy):
         Returns:
             Tuple[bool, str]: A boolean indicating correctness and the generated observation.
         """
+        _, execution_status = safe_execute(f"{query}\n\n{key}")
+
         self._scratchpad += f"\nObservation: "
         if action_type.lower() == "finish":
             self._finished = True
             self._answer = query
-            if EM(self._answer, key):
+            if EM(execution_status, "Done", normalize=False):
                 obs = "Answer is CORRECT"
             else:
                 obs = "Answer is INCORRECT"
         else:
-            obs = "Invalid action type, please try again."
+            obs = "Invalid action type, please try again. Valid action is Finish[```python<code>```]"
+
         self._scratchpad += obs
 
-        return EM(self._answer, key), obs
+        return EM(execution_status, "Done", normalize=False), obs
 
     def create_output_dict(
         self,
@@ -215,12 +251,7 @@ class ReflexionCoTQAStrategy(ReflexionCoTBaseStrategy):
             "reflections": reflections,
         }
 
-    def halting_condition(
-        self,
-        idx: int,
-        key: str,
-        **kwargs: Any,
-    ) -> bool:
+    def halting_condition(self, idx: int, key: str, **kwargs: Any) -> bool:
         """Determines whether the halting condition has been met.
 
         Args:
@@ -232,7 +263,8 @@ class ReflexionCoTQAStrategy(ReflexionCoTBaseStrategy):
             bool: True if the halting condition is met, False otherwise.
         """
         max_trials = kwargs.get("max_trials", self.max_trials)
-        return EM(self._answer, key) or idx >= max_trials
+        _, execution_status = safe_execute(f"{self._answer}\n\n{key}")
+        return EM(execution_status, "Done", normalize=False) or idx >= max_trials
 
     def reset(self, **kwargs: Any) -> None:
         """Resets the internal state of the strategy.
@@ -283,10 +315,7 @@ class ReflexionCoTQAStrategy(ReflexionCoTBaseStrategy):
         return reflections, reflections_str
 
     def reflect_condition(
-        self,
-        idx: int,
-        reflect_strategy: Optional[str],
-        key: str,
+        self, idx: int, reflect_strategy: Optional[str], key: str
     ) -> bool:
         """Determines whether the reflection condition has been met.
 
@@ -298,11 +327,16 @@ class ReflexionCoTQAStrategy(ReflexionCoTBaseStrategy):
         Returns:
             bool: True if the reflection condition is met, False otherwise.
         """
-        return idx > 0 and not EM(self._answer, key) and reflect_strategy is not None
+        _, execution_status = safe_execute(f"{self._answer}\n\n{key}")
+        return (
+            idx > 0
+            and not EM(execution_status, "Done", normalize=False)
+            and reflect_strategy is not None
+        )
 
 
-class ReflexionReActQAStrategy(ReflexionReActBaseStrategy):
-    """A strategy class for QA benchmarks using the ReflexionReAct agent.
+class ReflexionReActCodeStrategy(ReflexionReActBaseStrategy):
+    """A strategy class for Code benchmarks using the ReflexionReAct agent.
 
     Attributes:
         llm (BaseChatModel): The language model used for generating answers and critiques.
@@ -311,7 +345,6 @@ class ReflexionReActQAStrategy(ReflexionReActBaseStrategy):
         max_trials (int): The maximum number of trials allowed. Defaults to 1.
         max_steps (int): The maximum number of steps allowed. Defaults to 6.
         max_tokens (int): The maximum number of tokens allowed. Defaults to 5000.
-        docstore (DocstoreExplorer): The document store explorer for retrieving relevant documents. Defaults to Wikipedia.
         enc (Encoding): The encoding for tokenization. Defaults to gpt-3.5-turbo.
     """
 
@@ -323,24 +356,21 @@ class ReflexionReActQAStrategy(ReflexionReActBaseStrategy):
         max_trials: int = 1,
         max_steps: int = 6,
         max_tokens: int = 5000,
-        docstore: DocstoreExplorer = DocstoreExplorer(Wikipedia()),
         enc: Encoding = tiktoken.encoding_for_model("gpt-3.5-turbo"),
     ) -> None:
         """Initialization."""
         super().__init__(llm)
         self.max_reflections = max_reflections
         self.max_trials = max_trials
+        self.max_steps = max_steps
+        self.max_tokens = max_tokens
+        self.enc = enc
 
         if not reflector:
             reflector = ReflexionReActReflector(
                 llm=llm, max_reflections=max_reflections
             )
         self.reflector = reflector
-
-        self.max_steps = max_steps
-        self.max_tokens = max_tokens
-        self.docstore = docstore
-        self.enc = enc
 
         self._finished = False
         self._answer = ""
@@ -420,18 +450,15 @@ class ReflexionReActQAStrategy(ReflexionReActBaseStrategy):
             prompt=prompt,
             additional_keys=additional_keys,
         )
-        action = remove_newline(action).split("Observation")[0]
-        self._scratchpad += " " + action
-        action_type, query = parse_qa_action(action)
+        action = action.split("Observation")[0].strip()
+
+        action_type, query = parse_code_action_react(action)
+        self._scratchpad += f" {action_type}[\n```python\n{query}\n```\n]"
 
         return action_type, query
 
     def generate_observation(
-        self,
-        step_idx: int,
-        action_type: str,
-        query: str,
-        key: str,
+        self, step_idx: int, action_type: str, query: str, key: str
     ) -> Tuple[bool, str, Dict[str, Any]]:
         """Generate an observation based on the action type and query.
 
@@ -445,40 +472,45 @@ class ReflexionReActQAStrategy(ReflexionReActBaseStrategy):
             Tuple[bool, str, Dict[str, Any]]: A tuple containing a boolean indicating whether the answer is correct, a string representing the observation,
                 and a dictionary of the external tool outputs.
         """
-        external_tool_info = {"search_result": "", "lookup_result": ""}
+        external_tool_info = {"execution_status": ""}
 
         self._scratchpad += f"\nObservation {step_idx}: "
         if action_type.lower() == "finish":
+            obs = f"{query}\n\n{key}"
+            _, execution_status = safe_execute(obs)
+            external_tool_info["execution_status"] = execution_status
+
             self._answer = query
             self._finished = True
-            if EM(self._answer, key):
+
+            if EM(execution_status, "Done", normalize=False):
                 obs = "Answer is CORRECT"
             else:
                 obs = "Answer is INCORRECT"
-        elif action_type.lower() == "search":
-            try:
-                search_result = self.docstore.search(query)
-                external_tool_info["search_result"] = search_result
-                obs = remove_newline(search_result)
-            except Exception:
-                obs = "Could not find that page, please try again."
-        elif action_type.lower() == "lookup":
-            try:
-                lookup_result = self.docstore.lookup(query)
-                external_tool_info["lookup_result"] = lookup_result
-                obs = remove_newline(lookup_result)
-            except ValueError:
-                obs = "The last page Searched was not found, so you cannot Lookup a keyword in it. Please try one of the similar pages given."
+        elif action_type.lower() == "implement":
+            _, execution_status = safe_execute(query)
+            external_tool_info["execution_status"] = execution_status
+            execution_status = (
+                ""  # Execution status may be done, but not necessarily correct.
+            )
+
+            self._answer = query
+            obs = f"\n```python\n{self._answer}\n```\nExecution Status: {execution_status}"
+        elif action_type.lower() == "test":
+            obs = f"{self._answer}\n\n{query}"
+            _, execution_status = safe_execute(obs)
+            external_tool_info["execution_status"] = execution_status
+
+            obs = f"\n```python\n{obs}\n```\nExecution Status: {execution_status}"
         else:
-            obs = "Invalid Action. Valid Actions are Lookup[<topic>] Search[<topic>] and Finish[<answer>]."
+            execution_status = ""
+            obs = "Invalid Action. Valid Actions are Implement[code] Test[code] and Finish[answer]."
         self._scratchpad += obs
 
-        return EM(self._answer, key), obs, external_tool_info
+        return EM(execution_status, "Done", normalize=False), obs, external_tool_info
 
     def create_output_dict(
-        self,
-        react_out: List[ReflexionReActStepOutput],
-        reflections: List[str],
+        self, react_out: List[ReflexionReActStepOutput], reflections: List[str]
     ) -> Dict[str, Any]:
         """Create a dictionary containing the output of the ReflexionReAct agent.
 
@@ -538,7 +570,8 @@ class ReflexionReActQAStrategy(ReflexionReActBaseStrategy):
             bool: True if the halting condition is met, False otherwise. The halting condition is met when the answer is not correct and the current step index is less than the maximum number of trials plus one.
         """
         max_trials: int = kwargs.get("max_trials", self.max_trials)
-        return EM(self._answer, key) or idx >= max_trials + 1
+        _, execution_status = safe_execute(f"{self._answer}\n\n{key}")
+        return EM(execution_status, "Done", normalize=False) or idx >= max_trials + 1
 
     def react_halting_condition(
         self,
@@ -671,52 +704,74 @@ class ReflexionReActQAStrategy(ReflexionReActBaseStrategy):
             additional_keys=additional_keys,
         )
 
-        return halted and not EM(self._answer, key) and reflect_strategy is not None
+        _, execution_status = safe_execute(f"{self._answer}\n\n{key}")
+
+        return (
+            halted
+            and not EM(execution_status, "Done", normalize=False)
+            and reflect_strategy is not None
+        )
 
 
-class ReflexionCoTHotQAStrategy(ReflexionCoTQAStrategy):
-    """A strategy class for the HotpotQA benchmark using the ReflexionCoT agent."""
+class ReflexionCoTHEvalStrategy(ReflexionCoTCodeStrategy):
+    """A strategy class for the HumanEval benchmark using the ReflexionCoT agent."""
+
+    def generate_action(
+        self,
+        question: str,
+        examples: str,
+        reflections: str,
+        prompt: str,
+        additional_keys: Dict[str, str],
+        **kwargs: Any,
+    ) -> Tuple[str, str]:
+        """Generates an action based on the question, examples, and prompt.
+
+        Fixes the action_type to "Finish".
+
+        Args:
+            question (str): The question to be answered.
+            examples (str): Examples to guide the generation process.
+            reflections (str): Reflections to consider during generation.
+            prompt (str): The prompt used for generating the action.
+            additional_keys (Dict[str, str]): Additional keys for the generation process.
+            **kwargs (Any): Additional arguments.
+
+        Returns:
+            Tuple[str, str]: The generated action type and query.
+        """
+        self._scratchpad += "\nAction:"
+        action = _prompt_cot_agent(
+            llm=self.llm,
+            examples=examples,
+            reflections=reflections,
+            question=question,
+            scratchpad=self._scratchpad,
+            prompt=prompt,
+            additional_keys=additional_keys,
+        )
+        action = action.split("Observation")[0].strip()
+
+        query = action.split("```python")[-1].split("```")[0]
+        action_type = "Finish"
+        self._scratchpad += f" {action_type}[\n```python\n{query}\n```\n]"
+
+        return action_type, query
+
+
+class ReflexionCoTMBPPStrategy(ReflexionCoTCodeStrategy):
+    """A strategy class for the MBPP benchmark using the ReflexionCoT agent."""
 
     pass
 
 
-class ReflexionCoTTriviaQAStrategy(ReflexionCoTQAStrategy):
-    """A strategy class for the TriviaQA benchmark using the ReflexionCoT agent."""
+class ReflexionReActHEvalStrategy(ReflexionReActCodeStrategy):
+    """A strategy class for the HumanEval benchmark using the ReflexionReAct agent."""
 
     pass
 
 
-class ReflexionCoTAmbigNQStrategy(ReflexionCoTQAStrategy):
-    """A strategy class for the AmbigNQ benchmark using the ReflexionCoT agent."""
-
-    pass
-
-
-class ReflexionCoTFEVERStrategy(ReflexionCoTQAStrategy):
-    """A strategy class for the FEVER benchmark using the ReflexionCoT agent."""
-
-    pass
-
-
-class ReflexionReActHotQAStrategy(ReflexionReActQAStrategy):
-    """A strategy class for the HotpotQA benchmark using the ReflexionReAct agent."""
-
-    pass
-
-
-class ReflexionReActTriviaQAStrategy(ReflexionReActQAStrategy):
-    """A strategy class for the TriviaQA benchmark using the ReflexionReAct agent."""
-
-    pass
-
-
-class ReflexionReActAmbigNQStrategy(ReflexionReActQAStrategy):
-    """A strategy class for the AmbigNQ benchmark using the ReflexionReAct agent."""
-
-    pass
-
-
-class ReflexionReActFEVERStrategy(ReflexionReActQAStrategy):
-    """A strategy class for the FEVER benchmark using the ReflexionReAct agent."""
+class ReflexionReActMBPPStrategy(ReflexionReActCodeStrategy):
+    """A strategy class for the MBPP benchmark using the ReflexionReAct agent."""
 
     pass
