+++ agential/cog/strategies/strategy_factory.py
@@ -31,17 +31,17 @@ from agential.cog.react.strategies.qa import (
     ReActHotQAStrategy,
     ReActTriviaQAStrategy,
 )
-from agential.cog.strategies.reflexion.base import (
+from agential.cog.reflexion.strategies.base import (
     ReflexionCoTBaseStrategy,
     ReflexionReActBaseStrategy,
 )
-from agential.cog.strategies.reflexion.code import (
+from agential.cog.reflexion.strategies.code import (
     ReflexionCoTHEvalStrategy,
     ReflexionCoTMBPPStrategy,
     ReflexionReActHEvalStrategy,
     ReflexionReActMBPPStrategy,
 )
-from agential.cog.strategies.reflexion.math import (
+from agential.cog.reflexion.strategies.math import (
     ReflexionCoTGSM8KStrategy,
     ReflexionCoTSVAMPStrategy,
     ReflexionCoTTabMWPStrategy,
@@ -49,7 +49,7 @@ from agential.cog.strategies.reflexion.math import (
     ReflexionReActSVAMPStrategy,
     ReflexionReActTabMWPStrategy,
 )
-from agential.cog.strategies.reflexion.qa import (
+from agential.cog.reflexion.strategies.qa import (
     ReflexionCoTAmbigNQStrategy,
     ReflexionCoTFEVERStrategy,
     ReflexionCoTHotQAStrategy,
